fix(docs): [grid] update custom style docs
