+++ examples/sites/demos/pc/app/grid/custom-style/footer-style.vue
@@ -1,19 +1,38 @@
 <template>
-  <tiny-grid
-    class="grid-footer-cell-style"
-    :data="tableData"
-    show-footer
-    :footer-method="footerMethod"
-    border
-    :footer-cell-class-name="footerCellClassName"
-  >
-    <tiny-grid-column type="index" width="60"></tiny-grid-column>
-    <tiny-grid-column type="selection" width="60"></tiny-grid-column>
-    <tiny-grid-column field="name" title="公司名称"></tiny-grid-column>
-    <tiny-grid-column field="employees" title="员工数"></tiny-grid-column>
-    <tiny-grid-column field="createdDate" title="创建日期"></tiny-grid-column>
-    <tiny-grid-column field="city" title="城市"></tiny-grid-column>
-  </tiny-grid>
+  <div>
+    <h4>自定义表尾行样式：</h4>
+    <tiny-grid
+      class="grid-footer-row-style"
+      :data="tableData"
+      show-footer
+      :footer-method="footerMethod"
+      border
+      :footer-row-class-name="footerRowClassName"
+    >
+      <tiny-grid-column type="index" width="60"></tiny-grid-column>
+      <tiny-grid-column type="selection" width="60"></tiny-grid-column>
+      <tiny-grid-column field="name" title="公司名称"></tiny-grid-column>
+      <tiny-grid-column field="employees" title="员工数"></tiny-grid-column>
+      <tiny-grid-column field="createdDate" title="创建日期"></tiny-grid-column>
+      <tiny-grid-column field="city" title="城市"></tiny-grid-column>
+    </tiny-grid>
+    <h4>自定义表尾单元格样式：</h4>
+    <tiny-grid
+      class="grid-footer-cell-style"
+      :data="tableData"
+      show-footer
+      :footer-method="footerMethod"
+      border
+      :footer-cell-class-name="footerCellClassName"
+    >
+      <tiny-grid-column type="index" width="60"></tiny-grid-column>
+      <tiny-grid-column type="selection" width="60"></tiny-grid-column>
+      <tiny-grid-column field="name" title="公司名称"></tiny-grid-column>
+      <tiny-grid-column field="employees" title="员工数"></tiny-grid-column>
+      <tiny-grid-column field="createdDate" title="创建日期"></tiny-grid-column>
+      <tiny-grid-column field="city" title="城市"></tiny-grid-column>
+    </tiny-grid>
+  </div>
 </template>
 
 <script lang="jsx">
@@ -113,9 +132,18 @@ export default {
         })
       ]
     },
-    footerCellClassName({ $rowIndex }) {
-      if ($rowIndex === 0) {
+    footerCellClassName({ column, $rowIndex }) {
+      if ($rowIndex === 0 && column.property === 'employees') {
         return 'footer__cell--blue'
+      }
+    },
+    footerRowClassName({ $rowIndex }) {
+      if ($rowIndex === 0) {
+        return 'footer__row--red'
+      }
+
+      if ($rowIndex === 1) {
+        return 'footer__row--green'
       }
     }
   }
@@ -125,6 +153,15 @@ export default {
 <style>
 .grid-footer-cell-style .tiny-grid-footer__column.footer__cell--blue {
   background-color: #2db7f5;
+  color: #fff;
+}
+
+.grid-footer-row-style .tiny-grid-footer__row.footer__row--red {
+  background-color: palevioletred;
+  color: #fff;
+}
+.grid-footer-row-style .tiny-grid-footer__row.footer__row--green {
+  background-color: green;
   color: #fff;
 }
 </style>
