+++ examples/sites/demos/pc/app/grid/footer/configuration-summary-composition-api.vue
@@ -9,7 +9,7 @@
   </tiny-grid>
 </template>
 
-<script setup lang="jsx">
+<script setup>
 import { ref } from 'vue'
 import { Grid as TinyGrid, GridColumn as TinyGridColumn } from '@opentiny/vue'
