--- examples/sites/demos/pc/app/grid/footer/footer-summation.spec.js
@@ -1,7 +1,0 @@
-import { test, expect } from '@playwright/test'
-
-test('表尾统计', async ({ page }) => {
-  page.on('pageerror', (exception) => expect(exception).toBeNull())
-  await page.goto('grid-footer#footer-footer-summation')
-  await expect(page.getByRole('cell', { name: '5310' })).toBeVisible()
-})
