--- examples/sites/demos/pc/app/grid/summary/custom-summary.spec.js
@@ -1,7 +1,0 @@
-import { test, expect } from '@playwright/test'
-
-test('test', async ({ page }) => {
-  page.on('pageerror', (exception) => expect(exception).toBeNull())
-  await page.goto('grid-summary#summary-custom-summary')
-  await expect(page.getByText('5310')).toBeVisible()
-})
