+++ examples/sites/demos/pc/app/grid/webdoc/grid-footer.js
@@ -3,27 +3,22 @@ export default {
   owner: '',
   demos: [
     {
-      'demoId': 'footer-footer-summation',
-      'name': { 'zh-CN': '表尾统计', 'en-US': 'Table tail statistics' },
-      'desc': { 'zh-CN': '详细用法参考如下示例', 'en-US': 'For details, see the following example.' },
-      'codeFiles': ['footer/footer-summation.vue']
-    },
-    {
       'demoId': 'footer-footer-summation-empty',
       'name': { 'zh-CN': '表尾统计（空数据）', 'en-US': 'Table tail statistics (empty data)' },
-      'desc': { 'zh-CN': '详细用法参考如下示例', 'en-US': 'For details, see the following example.' },
+      'desc': {
+        'zh-CN': '通过 <code>footer-method</code> 和 <code>show-footer</code> 配置自定义统计。',
+        'en-US': 'For details, see the following example.'
+      },
       'codeFiles': ['footer/footer-summation-empty.vue']
     },
     {
-      'demoId': 'footer-footer-class-name',
-      'name': { 'zh-CN': '表尾class类名', 'en-US': 'Table tail class name' },
+      'demoId': 'footer-configuration-summary',
+      'name': { 'zh-CN': '设置配置式的统计', 'en-US': 'Set Profile Statistics' },
       'desc': {
-        'zh-CN':
-          '<p>\n通过表格列属性 <code>footer-class-name</code> 给表尾的单元格附加 className，也可以是函数 Function({$rowIndex, column, columnIndex, $columnIndex})。\n</p>\n',
-        'en-US':
-          '<p>\nAdd a class name to the cell at the end of the table through the table column attribute <code>footer-class-name</code> or the function Function({$rowIndex, column, columnIndex, $columnIndex}) . \n</p>\n'
+        'zh-CN': '<p>通过 <code>summary-config</code> 设置配置式的统计。</p>\n',
+        'en-US': '<p>Set the profile statistics using <code>summary-config</code>. </p>\n'
       },
-      'codeFiles': ['footer/footer-class-name.vue']
+      'codeFiles': ['footer/configuration-summary.vue']
     },
     {
       'demoId': 'footer-footer-row-or-column-span',
