--- examples/sites/demos/pc/app/grid/webdoc/grid-summary.en.md
@@ -1,7 +1,0 @@
----
-title: Grid Table
----
-
-# Grid Table
-
-<div>Table component, which provides powerful data table functions. In Grid, data lists can be displayed, selected, and edited.</div>
