--- examples/sites/demos/pc/app/grid/webdoc/grid-summary.json.cn.md
@@ -1,7 +1,0 @@
----
-title:  Grid 表格
----
-
-#  Grid 表格
-
-<div>表格组件，提供了非常强大数据表格功能，在 Grid 可以展示数据列表，可以对数据列表进行选择、编辑等。</div>
